split TestOngoingCachingRegister into a standalone module.
